pyproject.toml
[project]
name = "kueri"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=1.3.0",
    "crewai-tools>=1.3.0",
    "openai>=1.109.1",
    "psycopg2-binary",
    "streamlit>=1.51.0",
    "typing>=3.10.0.0",
    "python-dotenv>=1.2.1",
    "mcp>=1.20.0",
    "fastapi>=0.121.0",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.44",
    "pymysql>=1.1.2",
    "loguru>=0.7.3",
    "asyncpg>=0.30.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "langchain-mcp>=0.2.1",
]
